pyproject.toml
[tool.poetry]
name = "delivery-rating-provider"
version = "0.1.0"
description = ""
authors = ["Torben Carstens <git@carstens.tech>"]
readme = "README.md"
packages = [{ include = "provider", from = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
httpx = ">=0.25.2,<0.26.0"
python-telegram-bot = "^20.7"
aiocache = "^0.12.2"

[tool.poetry.group.dev.dependencies]
mypy = "^1.7.1"
black = "^23.12.1"
ruff = "^0.1.12"
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.mypy.overrides]]
module = "aiocache"
ignore_missing_imports = true

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
